changes following review

+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_foast_pretty_printer.py
@@ -17,7 +17,7 @@ import textwrap
 
 import pytest
 
-from gt4py.next import Dimension, DimensionKind, Field, field_operator, int64, scan_operator
+from gt4py.next import Dimension, DimensionKind, Field, field_operator, int32, int64, scan_operator
 from gt4py.next.ffront.ast_passes import single_static_assign as ssa
 from gt4py.next.ffront.foast_pretty_printer import pretty_format
 from gt4py.next.ffront.func_to_foast import FieldOperatorParser
@@ -82,8 +82,8 @@ def test_scanop():
 
     expected = textwrap.dedent(
         f"""
-        @scan_operator(axis=Dimension(value="KDim", kind=DimensionKind.VERTICAL), forward=False, init=1.)
-        def scan(inp: int64) -> int64:
+        @scan_operator(axis=Dimension(value="KDim", kind=DimensionKind.VERTICAL), forward=False, init=1)
+        def scan(inp: int32) -> int32:
           {ssa.unique_name("foo", 0)} = inp
           return inp
         """
